test: Migrated bluebird versioned tests to `node:test`
